% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{changepoint_dissimilarity}
\alias{changepoint_dissimilarity}
\title{Change in community composition across changepoints}
\usage{
changepoint_dissimilarity(modal_ts_preds)
}
\arguments{
\item{modal_ts_preds}{dataframe of species abundances (or proportions) with columns \code{year} and \code{seg}. Result of \code{modal_ts()}}
}
\value{
dataframe of \code{cpt}, \code{seg_before}, \code{seg_after}, and \code{dissimilarity}
}
\description{
Bray-Curtis dissimilarity comparing the community composition pre and post each changepoint.
}
