% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{find_modal_cpts}
\alias{find_modal_cpts}
\title{Find modal combined location of changepoints}
\usage{
find_modal_cpts(model_fit)
}
\arguments{
\item{model_fit}{list with \verb{$ts_mod}}
}
\value{
vector of changepoint locations (years)
}
\description{
Find most common JOINT combination of changepoint locations from a TS fit.
}
