% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_run_helpers.R
\name{fit_means_lda}
\alias{fit_means_lda}
\title{Make a toy LDA with species means}
\usage{
fit_means_lda(dataset, lda_seed)
}
\arguments{
\item{dataset}{either a full or subsetted dataset}

\item{lda_seed}{seed (shouldn't matter)}
}
\value{
LDA with modified beta and gammas
}
\description{
Make a toy LDA with species means
}
