% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplevel_run.R
\name{fit_multiple_ldats_crossval}
\alias{fit_multiple_ldats_crossval}
\title{Fit multiple LDATS crossval models}
\usage{
fit_multiple_ldats_crossval(
  dataset,
  buffer = 2,
  ks,
  lda_seeds,
  cpts,
  nit = 100,
  cpt_seeds = NULL,
  summarize_ll = T
)
}
\arguments{
\item{dataset}{dataset}

\item{buffer}{default 2}

\item{ks}{vector of integers, numbers of topics}

\item{lda_seeds}{vector of integers, LDA seeds}

\item{cpts}{vector of integers, cpts to try}

\item{nit}{integer number of iterations}

\item{cpt_seeds}{default NULL, or you can pass integers to use as the cpt seed}

\item{summarize_ll}{return summarized (TRUE), or for every year?}
}
\value{
data frame of model fits for all configuations
}
\description{
Wrapper function for running \code{fit_ldats_crossval} with multiple configurations. NOT RECOMMENDED for running at scale - useful for debugging on a local machine. Use a cluster to run at scale.
}
