% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_run_helpers.R
\name{get_abund_probabilities}
\alias{get_abund_probabilities}
\title{Get abundance probabilities}
\usage{
get_abund_probabilities(
  subsetted_dataset_item,
  fitted_lda,
  fitted_ts,
  max_sims = NULL
)
}
\arguments{
\item{subsetted_dataset_item}{list with (at least) element \verb{$full}: unaltered dataset.}

\item{fitted_lda}{LDA fit to either a subsetted or a full dataset}

\item{fitted_ts}{TS fit to that LDA}

\item{max_sims}{defaults NULL. If supplied, will randomly draw \code{max_sims} draws from the posterior and return probabilities just for those. Useful for speeding things up.}
}
\value{
list of matrices of abundance probabilities for each species (columns) for each timestep (rows), with one list for each draw from the posterior (or \code{max_sims}, if supplied).
}
\description{
Get probabilities of abundances, as the multinomial probabilities, for each species at each time step as predicted by a TS model.
}
\details{
Gets one matrix of probabiltiies for each draw from the posterior for estimates of changepoint locations and model parameters (in this case just the intercept).
}
