% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_short_portal.R
\name{get_rodents_annual}
\alias{get_rodents_annual}
\title{Get Portal rodent data annually}
\usage{
get_rodents_annual()
}
\value{
list of abundance and covariates in \code{MATSS} format
}
\description{
Gets Portal control rodents, summarized as annual sums.
}
