% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_run_helpers.R
\name{get_test_loglik}
\alias{get_test_loglik}
\title{Get test loglikelihood (all)}
\usage{
get_test_loglik(subsetted_dataset_item, abund_probabilities)
}
\arguments{
\item{subsetted_dataset_item}{list with - at least - elements \code{test} (of abundance matrix and covaraites for test timestep) and \code{test_timestep} (the row number, of the full dataset, of the test timestep).}

\item{abund_probabilities}{list of abund_probabilities; with an element for every draw from the posterior being considered.}
}
\value{
vector of loglikelihood of test data given every abund_probability estimate
}
\description{
Wrapper for \code{get_one_test_loglik.} Gets loglikelihood estimate of the test timestep for every draw from the posterior (or every draw supplied via \code{abund_probabilities}).
}
