% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_run.R
\name{ldats_subset_one}
\alias{ldats_subset_one}
\title{Run LDATS on a single dataset subset}
\usage{
ldats_subset_one(
  subsetted_dataset_item,
  k,
  seed,
  cpts,
  nit,
  return_full = FALSE
)
}
\arguments{
\item{subsetted_dataset_item}{result of subset_data_one}

\item{k}{ntopics for lda}

\item{seed}{seed for lda. only use even numbers.}

\item{cpts}{how many changepoints for ts?}

\item{nit}{how many iterations? (draws from posterior)}

\item{return_full}{return fitted model objects and abundance probabilities in addition to logliks. Can be useful for diagnostics, but hogs memory.}
}
\value{
list. subsetted_dataset_item with the following appended: If \code{return_full}, fitted_lda; fitted_ts; abund_probabilities, otherwise NULL; test_logliks, model_info
}
\description{
This function runs on a single subset (e.g. the dataset with timestep 1 as the test timestep). Run \code{fit_ldats_crossval} to run this function on every subset.
}
\details{
First, fits an LDA to the \emph{full} (not subsetted) dataset. Then splits the matrix of topic proportions (\code{gamma} matrix) for that LDA into training/test subsets to match the subset. (The LDA is fit to the full dataset, because LDAs fit to different subsets cannot be recombined in a logical way).

Then fits a TS model to the \emph{subsetted} \code{gamma} matrix, with the specified number of iterations & changepoints.

Then extracts from that TS model the predicted abundances (multinomial probability distribution of species abundances) for each timestep. Because of the Bayesian component of the changepoint model, there is a matrix of predicted abundances per timestep \emph{for every draw from the posterior}, so \code{nit} matrices. Then calculates the loglikelihood of the test timestep given these predicted probabilities. There are \code{nit} estimates of the loglikelihood.

Returns the subsetted dataset item list provided, with the following appendend: The LDA, TS, and abundance probabilities (if \code{return_full = TRUE}), or as NULL otherwise; the vector of loglikelihoods for the test timestep for each iteration; a list \code{model_info} with the model specifications \verb{(k, seed, cpts, nit)}
}
