% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{modal_ts}
\alias{modal_ts}
\title{Predictions from TS model using modal changepoint estimates}
\usage{
modal_ts(model_fit, type = "species_proportions")
}
\arguments{
\item{model_fit}{list with elements \verb{$dataset}, \verb{$lda_mod}, \verb{$ts_mod}}

\item{type}{"topic_proportions", "species_proportions", or "species_abundances"}
}
\value{
dataframe of predicted values by year and segment
}
\description{
Get predicted topic proportions, species proportions, or species abundances from the modal changepoint estimates from a TS model.
}
