% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{r2s}
\alias{r2s}
\title{R2s from TS fit}
\usage{
r2s(model_fit, use_proportional = TRUE)
}
\arguments{
\item{model_fit}{list with \code{dataset}, \code{lda_mod}, \code{ts_mod}}

\item{use_proportional}{whether to use proportional abundances (T) or actual counts (F)}
}
\value{
dataframe of model r2 and r2 of just fitting the mean for each species.
}
\description{
R2s from TS fit
}
