% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model.R
\name{run_best_model}
\alias{run_best_model}
\title{Run best-fitting model}
\usage{
run_best_model(dataset, selected_config, nit = 1000, cpt_seed = NULL)
}
\arguments{
\item{dataset}{dataset}

\item{selected_config}{result of select_cvlt; or dataframe with 1 row of \code{k}, \code{lda_seed}, \code{cpts}}

\item{nit}{default 1000, nit}

\item{cpt_seed}{optionally pass a seed to the cpt model, default NULL}
}
\value{
list with dataset, config, LDA, and TS models
}
\description{
Run the LDATS model matching the best configuration for a dataset
}
