% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/best_model.R
\name{select_cvlt}
\alias{select_cvlt}
\title{Select simplest good model}
\usage{
select_cvlt(cvlt_fits, nse = 1)
}
\arguments{
\item{cvlt_fits}{result of running fit_ldats_multiple}

\item{nse}{number of SEs to allow, default 1}
}
\value{
dataframe with specifications for simplest model with loglikelihood within 1 SE of the loglikelihood of the best model
}
\description{
Select simplest fit within one SE of best fit
}
\details{
First minimizes the number of changepoints, then the number of topics. Then selects the LDA seed with the best loglikelihood.
}
