% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data_all}
\alias{subset_data_all}
\title{Subset data - all subsets}
\usage{
subset_data_all(full_dataset, buffer_size = 2)
}
\arguments{
\item{full_dataset}{MATSS-style dataset. A list with elements \verb{$abundance}, \verb{$covariates}}

\item{buffer_size}{number of timesteps to withhold on either side of the test timestep. Defaults 2.}
}
\value{
a list of ntimesteps lists, if ntimesteps is the number of timesteps in \code{full_dataset}. Each list in the list is the output of \code{subset_data_one} using one of the timesteps in the full dataset as the test timestep.
}
\description{
Wrapper for \code{subset_data_one} to create a list of all data subsets using one timestep of test data per subset and a buffer on either side.
}
