% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_data.R
\name{subset_data_one}
\alias{subset_data_one}
\title{Subset data - one subset}
\usage{
subset_data_one(full_dataset, test_timestep, buffer_size)
}
\arguments{
\item{full_dataset}{MATSS style dataset. A list with elements \verb{$abundance}, \verb{$covariates}}

\item{test_timestep}{Which timestep (row) to withold and use for test}

\item{buffer_size}{How many rows to withold on either side (not used for test)}
}
\value{
list with elements train (list of $abundance, $covariates), test (list of $abundance, $covariates), full (unaltered full_dataset), test_timestep (which tstep is the test one), buffer_size (buffer size)
}
\description{
Creates a \emph{single} training/test subset for a dataset.
Witholds a single timestep for testing, and a buffer of timesteps around that timestep.
}
