% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_run_helpers.R
\name{subset_lda}
\alias{subset_lda}
\title{Subset an LDA model}
\usage{
subset_lda(fitted_lda, subsetted_dataset_item)
}
\arguments{
\item{fitted_lda}{LDA fit to FULL dataset}

\item{subsetted_dataset_item}{result of subset_data_one, list with \verb{$test} and \verb{$train} elements.}
}
\value{
fitted_lda subset to just the timesteps present in the \verb{$train} element of \code{subsetted_dataset_item}.
}
\description{
Takes an LDA model fit to an ENTIRE dataset and extracts just rows present in the TRAIN data for a subsetted data item.
}
\details{
Subsets both the topic proportions  and loglikelihood, to maintain the correct dimensions.
}
