% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset_run_helpers.R
\name{summarize_ldats_fit}
\alias{summarize_ldats_fit}
\title{Summarize LDATS fits}
\usage{
summarize_ldats_fit(ldats_fits, summarize_ll = TRUE)
}
\arguments{
\item{ldats_fits}{a list of fitted models, from running fit_ldats_crossval(return_fits = T)}

\item{summarize_ll}{logical, whether to summmarize or not.}
}
\value{
data frame
}
\description{
Summarize LDATS fits
}
