% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_summary.R
\name{summarize_model}
\alias{summarize_model}
\title{Summarize a model fit}
\usage{
summarize_model(model_fit)
}
\arguments{
\item{model_fit}{list of \verb{$dataset}, \verb{$lda_mod}, \verb{$ts_mod}}
}
\value{
dataframe with info about the lda and ts models
}
\description{
Extract summary info
}
