% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toplevel_run.R
\name{fit_ldats_crossval}
\alias{fit_ldats_crossval}
\title{Run LDATs on a dataset using crossvalidation}
\usage{
fit_ldats_crossval(dataset, buffer = 2, k, seed, cpts, nit)
}
\arguments{
\item{dataset}{MATSS-style dataset (list with \verb{$abundance} and \verb{$covariates})}

\item{buffer}{number of timesteps to withold on either side of the test timestep, default 2}

\item{k}{integer number of topics}

\item{seed}{integer seed to use to run LDA}

\item{cpts}{integer number of changepoints}

\item{nit}{integer number of iterations for the changepoint model. 100 is fast but will not find the global optimum, 1000 gets closer but takes time.}
}
\value{
list of lists. Each element is the result of running \code{ldats_subset_one} on one subset of the original dataset.
}
\description{
Top-level wrapper to subset a dataset and fit LDATS to the subsets.
}
