% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_short_portal.R
\name{get_plants_annual}
\alias{get_plants_annual}
\title{Get Portal plants}
\usage{
get_plants_annual(census_season = "winter")
}
\arguments{
\item{census_season}{"winter" or summmer}
}
\value{
ready for ldats
}
\description{
Get Portal plants
}
